requirements.txt
# Python Telegram Bot framework
python-telegram-bot>=20.0.0,<21.0.0
